Splitting responsibilites, extracted CapabilitiesModifiers and applying them in the DesiredCapabilitiesProvider
