+++ bb-core/src/test/java/com/cognifide/qa/bb/provider/selenium/webdriver/modifiers/WebDriverModifiersTest.java
@@ -22,17 +22,13 @@ package com.cognifide.qa.bb.provider.selenium.webdriver.modifiers;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
-import java.util.Collections;
-
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InOrder;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.openqa.selenium.Capabilities;
 import org.openqa.selenium.WebDriver;
 
-import com.cognifide.qa.bb.provider.selenium.webdriver.modifiers.capabilities.CapabilitiesModifier;
 import com.cognifide.qa.bb.provider.selenium.webdriver.modifiers.webdriver.WebDriverModifier;
 import com.google.common.collect.Sets;
 
@@ -51,19 +47,7 @@ public class WebDriverModifiersTest {
   private WebDriverModifier mod3;
 
   @Mock
-  private CapabilitiesModifier capMod1;
-
-  @Mock
-  private CapabilitiesModifier capMod2;
-
-  @Mock
-  private CapabilitiesModifier capMod3;
-
-  @Mock
   private WebDriver webDriver;
-
-  @Mock
-  private Capabilities capabilities;
 
   @Test
   public void shouldSortAndFilterWebDriverModifiers() {
@@ -79,20 +63,6 @@ public class WebDriverModifiersTest {
     verify(mod3, never()).modify(webDriver);
   }
 
-  @Test
-  public void shouldSortAndFilterCapabilitiesModifiers() {
-    //when
-    setupCapabilitiesModifiers();
-    testedObject.modifyCapabilities(capabilities);
-
-    //then
-    InOrder inOrder = inOrder(capMod2, capMod1);
-    inOrder.verify(capMod2).modify(capabilities);
-    inOrder.verify(capMod1).modify(capabilities);
-
-    verify(capMod3, never()).modify(capabilities);
-  }
-
   public void setupWebDriverModifiers() {
     when(mod1.getOrder()).thenReturn(1);
     when(mod1.shouldModify()).thenReturn(true);
@@ -103,19 +73,6 @@ public class WebDriverModifiersTest {
     when(mod3.shouldModify()).thenReturn(false);
 
     testedObject =
-        spy(new WebDriverModifiers(Collections.emptySet(), Sets.newHashSet(mod1, mod2, mod3)));
-  }
-
-  public void setupCapabilitiesModifiers() {
-    when(capMod1.getOrder()).thenReturn(1);
-    when(capMod1.shouldModify()).thenReturn(true);
-    when(capMod1.modify(any())).thenReturn(capabilities);
-    when(capMod2.getOrder()).thenReturn(-1);
-    when(capMod2.shouldModify()).thenReturn(true);
-    when(capMod2.modify(any())).thenReturn(capabilities);
-    when(capMod3.shouldModify()).thenReturn(false);
-
-    testedObject = spy(new WebDriverModifiers(Sets.newHashSet(capMod1, capMod2, capMod3),
-        Collections.emptySet()));
+        spy(new WebDriverModifiers(Sets.newHashSet(mod1, mod2, mod3)));
   }
 }
